Add files via upload
A program to display the Full month calendar or a single month of  the given year.I found this calculation and logic in many calendar programs in the internet and I write this program by referring that.Input validation is not done. In this the application waits for user choice once it displayed the calendar user can either exit or continue using the options.